CHORE: move routeManager to domain ADD: reset method to route manager
